requirements.txt
pandas~=2.0.2
docxtpl~=0.16.7
streamlit~=1.23.1
numpy~=1.24.3
plotly~=5.14.1
pytexit~=0.4.0
Django=4
linque~=5.0.0
Django~=4.2
django-debug-toolbar~=4.3.0
requests~=2.31.0
dacite~=1.8.1
djangorestframework~=3.14.0